Created feed page

See merge request fiit_coworking/e-library/e-library-client!15

+++ src/app/common/delete-dialog/delete-dialog.component.ts
@@ -24,7 +24,12 @@ export class DeleteDialogComponent implements OnInit {
   }
 
   onYesClcik(): void {
-    this.adminService.deleteEntry(this.data.entryApikey).subscribe();
+    if(this.data.source === "admin"){
+      this.adminService.deleteEntry(this.data.entryApikey).subscribe();
+    }
+    if(this.data.source === "feed"){
+      this.adminService.deleteFeed(this.data.entryApikey).subscribe();
+    }
     this.dialogRef.close();
   }
 }
